+++ src/components/Books.js
@@ -1,15 +1,12 @@
 import React, { useState } from 'react';
-import { useDispatch } from 'react-redux';
+import './Main.css';
+import './Book.css';
 import { CircularProgressbar, buildStyles } from 'react-circular-progressbar';
+import Header from './Header';
 import 'react-circular-progressbar/dist/styles.css';
-import Header from '../../components/Header';
-import '../../components/Book.css';
-import '../../components/Main.css';
-
-// end imports
 
 function Book({
-  title, author, progress, category, onProgressUpdate, onRemoveBook,
+  title, author, progress, category, onProgressUpdate,
 }) {
   return (
     <div className="bookCard">
@@ -35,7 +32,7 @@ function Book({
             </span>
 
             <ul className="list">
-              <li onClick={onRemoveBook} className="firstListItem">remove</li>
+              <li>remove</li>
               <li>edit</li>
               <li>delete</li>
             </ul>
@@ -59,6 +56,7 @@ function Book({
           <i>Current Chapter</i>
           <button type="button" onClick={onProgressUpdate}>Update Progress</button>
         </div>
+
       </div>
     </div>
   );
@@ -92,41 +90,24 @@ function AdditionForm() {
   const [author, setAuthor] = useState('');
   const [category, setCategory] = useState('');
   const [books, setBooks] = useLocalStorage('books', []);
-  const dispatch = useDispatch();
 
   const onSubmit = (e) => {
     e.preventDefault();
-    const newBook = {
-      title, author, category, progress: 0,
-    };
-    dispatch({
-      type: 'ADD_BOOK',
-      payload: newBook,
-    });
-    setBooks([...books, newBook]);
+    setBooks((prevBooks) => [
+      ...prevBooks,
+      {
+        title, author, category, progress: 0,
+      },
+    ]);
     setTitle('');
     setAuthor('');
     setCategory('');
   };
 
   const onUpdateProgressBarClick = (index) => {
-    dispatch({
-      type: 'UPDATE_PROGRESS',
-      payload: { index },
-    });
-    const updatedBooks = [...books];
-    updatedBooks[index].progress += 1;
-    setBooks(updatedBooks);
-  };
-
-  const onRemoveBookClick = (index) => {
-    dispatch({
-      type: 'REMOVE_BOOK',
-      payload: { index },
-    });
-    const updatedBooks = [...books];
-    updatedBooks.splice(index, 1);
-    setBooks(updatedBooks);
+    setBooks((prevBooks) => prevBooks.map((book, i) => (i === index
+      ? { ...book, progress: Math.min(book.progress + 5, 100) }
+      : book)));
   };
 
   return (
@@ -141,7 +122,6 @@ function AdditionForm() {
             progress={book.progress}
             category={book.category}
             onProgressUpdate={() => onUpdateProgressBarClick(index)}
-            onRemoveBook={() => onRemoveBookClick(index)}
           />
         ))}
       </div>
