+++ src/redux/store.js
@@ -1,27 +1,17 @@
-import { createStore } from 'redux';
+import { configureStore, combineReducers } from '@reduxjs/toolkit';
+import categoriesReducer from './categories/categoriesSlice';
+import booksReducer from './books/booksSlice';
 
-const initialState = {
-  books: [],
+// end immports
+
+const rootReducer = combineReducers({
+  categories: categoriesReducer,
+  books: booksReducer,
+});
+
+const storeConfig = () => {
+  const store = configureStore(rootReducer);
+  return store;
 };
 
-function booksReducer(state = initialState, action) {
-  switch (action.type) {
-    case 'ADD_BOOK':
-      return { ...state, books: [...state.books, action.payload] };
-    case 'REMOVE_BOOK':
-      return { ...state, books: state.books.filter((_, index) => index !== action.payload.index) };
-    case 'UPDATE_PROGRESS':
-      return {
-        ...state,
-        books: state.books.map((book, i) => (i === action.payload.index
-          ? { ...book, progress: Math.min(book.progress + 5, 100) }
-          : book)),
-      };
-    default:
-      return state;
-  }
-}
-
-const store = createStore(booksReducer);
-
-export default store;
+export default storeConfig;
